move page components to pages folder
