+++ src/pages/BookingConfirmation/Confirmation.scss
