+++ src/pages/CarDetails/CarDetails.scss
